# src/ehs/docs/openapi.yaml
openapi: 3.1.0
info:
  title: SpaceOS JoineryTech EHS API
  version: 1.0.0
  description: |
    Environmental Health & Safety (Munkavédelem/ISO 45001) API for the JoineryTech ERP system.

    ## Features
    - **Incident Management** — Accident, near-miss, and hazardous condition reporting with FSM workflow
    - **Risk Assessment** — 5×5 risk matrix for hazard evaluation and control measures
    - **Training/Competency Tracking** — Employee training records with expiration monitoring
    - **ISO 45001 Compliance** — Occupational Health & Safety management system support
    - **HR Integration** — Employee links for incident reporting and training records

    ## Authentication
    All endpoints require JWT Bearer token authentication. Tenant isolation is enforced via PostgreSQL RLS.

    ## Multi-tenant Architecture
    - TenantId extracted from JWT token
    - All queries automatically filtered by tenant
    - Cross-tenant access is not possible

    ## 5×5 Risk Matrix
    - Severity: 1 (Negligible) → 5 (Catastrophic)
    - Likelihood: 1 (Rare) → 5 (Almost Certain)
    - RiskScore = Severity × Likelihood (1-25)
    - RiskLevel: CONFIG-DRIVEN bands (Ehs:RiskMatrix:Bands) —
      defaults: Low 1-4, Medium 5-9, High 10-16, Critical 17-25
    - FSM: Draft → UnderReview → Approved → Archived (+ return-to-draft)

  contact:
    name: SpaceOS Development Team
    email: dev@spaceos.hu
  license:
    name: Proprietary
    url: https://spaceos.hu/license
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/ehs/ppe-items:
    get:
      operationId: listPpeItems
      tags: [PpeItems]
      summary: List PPE catalogue items
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PpeCategory'
      responses:
        '200':
          description: PPE catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PpeItemDto'

    post:
      operationId: createPpeItem
      tags: [PpeItems]
      summary: Create PPE catalogue item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePpeItemRequest'
      responses:
        '201':
          description: Item created (returns the new id)
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/ehs/ppe-items/{id}:
    get:
      operationId: getPpeItem
      tags: [PpeItems]
      summary: Get PPE catalogue item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PPE item detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PpeItemDto'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updatePpeItem
      tags: [PpeItems]
      summary: Update PPE catalogue item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePpeItemRequest'
      responses:
        '204':
          description: Item updated
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/ehs/ppe-items/{id}/deactivate:
    post:
      operationId: deactivatePpeItem
      tags: [PpeItems]
      summary: Soft-deactivate a PPE catalogue item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deactivated
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/ehs/ppe-issuances:
    get:
      operationId: listPpeIssuances
      tags: [PpeIssuances]
      summary: List PPE issuances
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PpeIssuanceStatus'
        - name: expiringWithinDays
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Issuance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PpeIssuanceDto'

    post:
      operationId: issuePpe
      tags: [PpeIssuances]
      summary: Record a PPE hand-out (FSM entry — Issued/kiadva)
      description: |
        When expiresAt is omitted it is derived from PpeItem.defaultLifetimeMonths.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuePpeRequest'
      responses:
        '201':
          description: Issuance created (returns the new id)
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/ehs/ppe-issuances/expiring:
    get:
      operationId: listExpiringPpeIssuances
      tags: [PpeIssuances]
      summary: Outstanding issuances expiring within the window (dashboard)
      parameters:
        - name: withinDays
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Expiring outstanding issuances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PpeIssuanceDto'
  /api/ehs/ppe-issuances/by-employee/{employeeId}:
    get:
      operationId: getEmployeePpeSheet
      tags: [PpeIssuances]
      summary: Employee PPE sheet (portal view)
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Every issuance of the employee
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PpeIssuanceDto'
  /api/ehs/ppe-issuances/{id}:
    get:
      operationId: getPpeIssuance
      tags: [PpeIssuances]
      summary: Get PPE issuance by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Issuance detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PpeIssuanceDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ehs/ppe-issuances/{id}/acknowledge:
    post:
      operationId: acknowledgePpeIssuance
      tags: [PpeIssuances]
      summary: "FSM: Issued → Acknowledged (atvett)"
      description: Illegal transition returns 409 with the domain guard message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Acknowledged
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/ehs/ppe-issuances/{id}/return:
    post:
      operationId: returnPpeIssuance
      tags: [PpeIssuances]
      summary: "FSM: Acknowledged → Returned (visszavett, terminal)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Returned
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/ehs/ppe-issuances/{id}/replace:
    post:
      operationId: replacePpeIssuance
      tags: [PpeIssuances]
      summary: "FSM: Acknowledged → Replaced (cserelve) — spawns a new issuance"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplacePpeIssuanceRequest'
      responses:
        '201':
          description: Replacement issuance created (returns the NEW issuance id)
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  # ========================================
  # SAFETY WALKS + UNIFIED CAPA (10 endpoints)
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: ['object', 'null']

    # ========================================
    # LOCATION REGISTRY
    # ========================================
...
    PpeCategory:
      type: string
      enum: [fej, szem, hallas, legzes, kez, lab, test, leeses]
    PpeIssuanceStatus:
      type: string
      enum: [kiadva, atveve, visszaadva, cserelve]
      description: "FSM: Issued(kiadva) → Acknowledged(atvett) → Returned(visszavett) | Replaced(cserelve)"
    PpeItemDto:
      type: object
      required: [ppeItemId, tenantId, name, category, isActive, createdAt]
      properties:
        ppeItemId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        category:
          $ref: '#/components/schemas/PpeCategory'
        standardRef:
          type: ['string', 'null']
          description: Reference standard, e.g. "EN 388"
        defaultLifetimeMonths:
          type: ['integer', 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreatePpeItemRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          maxLength: 200
        category:
          $ref: '#/components/schemas/PpeCategory'
        standardRef:
          type: ['string', 'null']
        defaultLifetimeMonths:
          type: ['integer', 'null']
          minimum: 1
    PpeIssuanceDto:
      type: object
      required: [issuanceId, tenantId, employeeId, ppeItemId, issuedAt, issuedBy, quantity, status, isExpired]
      properties:
        issuanceId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
          description: FK to HR module
        ppeItemId:
          type: string
          format: uuid
        issuedAt:
          type: string
          format: date-time
        issuedBy:
          type: string
          format: uuid
        quantity:
          type: integer
        expiresAt:
          type: ['string', 'null']
          format: date-time
        status:
          $ref: '#/components/schemas/PpeIssuanceStatus'
        acknowledgedAt:
          type: ['string', 'null']
          format: date-time
        returnedAt:
          type: ['string', 'null']
          format: date-time
        replacedAt:
          type: ['string', 'null']
          format: date-time
        replacementIssuanceId:
          type: ['string', 'null']
          format: uuid
          description: Set on Replaced — the new issuance
        isExpired:
          type: boolean
          description: CALCULATED — expiresAt passed while the PPE is still out
    IssuePpeRequest:
      type: object
      required: [employeeId, ppeItemId, issuedBy, quantity]
      properties:
        employeeId:
          type: string
          format: uuid
        ppeItemId:
          type: string
          format: uuid
        issuedBy:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        expiresAt:
          type: ['string', 'null']
          format: date-time
          description: Optional — derived from defaultLifetimeMonths when omitted
    ReplacePpeIssuanceRequest:
      type: object
      required: [replacedBy]
      properties:
        replacedBy:
          type: string
          format: uuid
        newExpiresAt:
          type: ['string', 'null']
          format: date-time

    # ========================================
    # SAFETY WALKS + UNIFIED CAPA
    # ========================================
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Domain guard violation (illegal FSM transition or business rule)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'